separate tests across resources, provide basic test
